board/radxa/rock5b: use kernel 6.6
The upstream kernel now supports the rock5b board quite well, so let's
migrate away from the downstream kernel which is stuck on a 5.10 kernel.

We still need some minor devicetree patches to add all the PCIe nodes,
however these are already in linux-next and will be part of linux v6.7.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>