configs/rock5b: enable eudev to enable automatic module loading
The PHY driver used by the PCIe Gen2 controllers is built as a module
and is therefore not available at early boot.

We could change this module to be built in, but that is for amateurs.
Instead enable eudev so that it can load drivers that were unavailable
at early boot.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>