board/radxa/rock5b: enable DW eDMA core
The DW PCI controller drivers (host and endpoint mode) are currently
always built-in (as they cannot be built as modules), so if the DW eDMA
core is built as a module (CONFIG_DW_EDMA=m), the eDMA controllers will
not be registered even if the dw-edma module is later loaded.

Thus, CONFIG_DW_EDMA always needs to be built as built in.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>