use git from local FS
this is much faster than cloning from a remote repo

NOTE: for this to work you need to have a local branch with the name
'endpoint-dmafixes'

e.g.
$ git remote add nks-github https://github.com/floatious/linux.git
$ git fetch nks-github
$ git co -b endpoint-dmafixes nks-github/endpoint-dmafixes

Note that if there are updates on the remote brach, you need to
update your local branch before rebuilding linux in buildroot.

Signed-off-by: Niklas Cassel <cassel@kernel.org>